+++ stream-chat-android-state/src/main/java/io/getstream/chat/android/state/plugin/listener/internal/ShuffleGiphyListenerState.kt
@@ -40,6 +40,7 @@ internal class ShuffleGiphyListenerState(private val logic: LogicRegistry) : Shu
         if (result is Result.Success) {
             val processedMessage = result.value.copy(syncStatus = SyncStatus.COMPLETED)
             logic.channelFromMessage(processedMessage)?.upsertMessage(processedMessage)
+            logic.threads().upsertMessage(processedMessage)
             logic.threadFromMessage(processedMessage)?.upsertMessage(processedMessage)
         }
     }