+++ stream-chat-android-state/src/main/java/io/getstream/chat/android/state/plugin/logic/querythreads/internal/QueryThreadsLogic.kt
@@ -14,19 +14,23 @@
  * limitations under the License.
  */
 
-package io.getstream.chat.android.state.plugin.logic.channel.thread.internal
+package io.getstream.chat.android.state.plugin.logic.querythreads.internal
 
 import io.getstream.chat.android.client.api.models.QueryThreadsRequest
 import io.getstream.chat.android.client.events.ChatEvent
 import io.getstream.chat.android.client.events.MessageDeletedEvent
+import io.getstream.chat.android.client.events.MessageReadEvent
 import io.getstream.chat.android.client.events.MessageUpdatedEvent
+import io.getstream.chat.android.client.events.NewMessageEvent
 import io.getstream.chat.android.client.events.NotificationChannelDeletedEvent
+import io.getstream.chat.android.client.events.NotificationMarkUnreadEvent
 import io.getstream.chat.android.client.events.NotificationThreadMessageNewEvent
 import io.getstream.chat.android.client.events.ReactionDeletedEvent
 import io.getstream.chat.android.client.events.ReactionNewEvent
 import io.getstream.chat.android.client.events.ReactionUpdateEvent
 import io.getstream.chat.android.models.Message
 import io.getstream.chat.android.models.QueryThreadsResult
+import io.getstream.chat.android.models.Thread
 import io.getstream.log.taggedLogger
 import io.getstream.result.Error
 import io.getstream.result.Result
@@ -35,10 +39,17 @@ import io.getstream.result.Result
  * Logic class for "Query Threads" operations.
  *
  * @param stateLogic The [QueryThreadsStateLogic] managing the global state of the threads list.
+ * @param databaseLogic The [QueryThreadsDatabaseLogic] retrieving the (optional) offline database data.
  */
-internal class QueryThreadsLogic(private val stateLogic: QueryThreadsStateLogic) {
+internal class QueryThreadsLogic(
+    private val stateLogic: QueryThreadsStateLogic,
+    private val databaseLogic: QueryThreadsDatabaseLogic,
+) {
 
     private val logger by taggedLogger("Chat:QueryThreadsLogic")
+
+    // Keeps track of the threads that we loaded from back-end in the current session
+    private var threadIdsInSession: MutableList<String> = mutableListOf()
 
     /**
      * Run precondition for the request. If it returns [Result.Success] then the request is run otherwise it returns
@@ -69,11 +80,18 @@ internal class QueryThreadsLogic(private val stateLogic: QueryThreadsStateLogic)
      *
      * @param request The [QueryThreadsRequest] used to fetch the threads.
      */
-    internal fun onQueryThreadsRequest(request: QueryThreadsRequest) {
-        if (request.isNextPageRequest()) {
+    internal suspend fun onQueryThreadsRequest(request: QueryThreadsRequest) {
+        val isNextPageRequest = request.isNextPageRequest()
+        if (isNextPageRequest) {
             stateLogic.setLoadingMore(true)
         } else {
             stateLogic.setLoading(true)
+        }
+        if (isForceReload(request)) {
+            stateLogic.clearThreads()
+            stateLogic.clearUnseenThreadIds()
+        } else if (!isNextPageRequest) {
+            queryThreadsOffline(request)
         }
     }
 
@@ -85,18 +103,22 @@ internal class QueryThreadsLogic(private val stateLogic: QueryThreadsStateLogic)
      * operation failed.
      * @param request The [QueryThreadsRequest] used to fetch the threads.
      */
-    internal fun onQueryThreadsResult(result: Result<QueryThreadsResult>, request: QueryThreadsRequest) {
+    internal suspend fun onQueryThreadsResult(result: Result<QueryThreadsResult>, request: QueryThreadsRequest) {
         stateLogic.setLoadingMore(false)
         stateLogic.setLoading(false)
         when (result) {
             is Result.Success -> {
                 if (request.isNextPageRequest()) {
-                    stateLogic.appendThreads(result.value.threads)
+                    stateLogic.upsertThreads(result.value.threads)
+                    threadIdsInSession += result.value.threads.map(Thread::parentMessageId)
                 } else {
                     stateLogic.setThreads(result.value.threads)
                     stateLogic.clearUnseenThreadIds()
+                    threadIdsInSession = result.value.threads.map(Thread::parentMessageId).toMutableList()
                 }
                 stateLogic.setNext(result.value.next)
+                // Update local threads order
+                databaseLogic.setLocalThreadsOrder(threadIdsInSession)
             }
 
             is Result.Failure -> {
@@ -112,12 +134,37 @@ internal class QueryThreadsLogic(private val stateLogic: QueryThreadsStateLogic)
      */
     internal fun handleEvents(events: List<ChatEvent>) = events.forEach(::handleEvent)
 
+    /**
+     * Retrieves a [Message] by its ID if it is stored in the Threads state.
+     */
+    internal fun getMessage(messageId: String): Message? =
+        stateLogic.getMessage(messageId)
+
+    /**
+     * Upsert the given [Message] in a [Thread] if such exists.
+     */
+    internal fun upsertMessage(message: Message) = updateParentOrReply(message)
+
+    /**
+     * Upsert the given [Message] from a [Thread] if such exists.
+     */
+    internal fun deleteMessage(message: Message) =
+        stateLogic.deleteMessage(message)
+
     private fun handleEvent(event: ChatEvent) {
         when (event) {
-            is NotificationThreadMessageNewEvent -> addNewThreadMessage(event)
+            // Destructive operation - remove the threads completely from the list
             is NotificationChannelDeletedEvent -> deleteThreadsFromChannel(event.cid)
-            is MessageDeletedEvent -> updateParentOrReply(event.message)
+            // Informs about a new thread (loaded, not loaded, or newly created thread)
+            is NotificationThreadMessageNewEvent -> onNewThreadMessageNotification(event)
+            // (Potentially) Informs about marking a thread as unread
+            is NotificationMarkUnreadEvent -> markThreadAsUnread(event)
+            // (Potentially) Informs about reading of a thread
+            is MessageReadEvent -> markThreadAsRead(event)
+            // (Potentially) Updates/Inserts a message in a thread
+            is NewMessageEvent -> updateParentOrReply(event.message)
             is MessageUpdatedEvent -> updateParentOrReply(event.message)
+            is MessageDeletedEvent -> updateParentOrReply(event.message)
             is ReactionNewEvent -> updateParentOrReply(event.message)
             is ReactionUpdateEvent -> updateParentOrReply(event.message)
             is ReactionDeletedEvent -> updateParentOrReply(event.message)
@@ -127,18 +174,17 @@ internal class QueryThreadsLogic(private val stateLogic: QueryThreadsStateLogic)
 
     private fun QueryThreadsRequest.isNextPageRequest() = this.next != null
 
-    private fun addNewThreadMessage(event: NotificationThreadMessageNewEvent) {
+    private fun isForceReload(request: QueryThreadsRequest) =
+        !request.isNextPageRequest() && stateLogic.getUnseenThreadIds().isNotEmpty()
+
+    private fun onNewThreadMessageNotification(event: NotificationThreadMessageNewEvent) {
+        val newMessageThreadId = event.message.parentId ?: return
+        // Update the unseenThreadIsd if the relevant thread is not loaded (yet)
         val threads = stateLogic.getThreads()
-        val thread = threads.find { it.parentMessageId == event.message.parentId }
-        if (thread == null) {
-            // Thread is not (yet) loaded, just update the state of unseenThreadIds
-            event.message.parentId?.let { parentId ->
-                stateLogic.addUnseenThreadId(parentId)
-            }
-            return
+        if (threads.none { it.parentMessageId == newMessageThreadId }) {
+            stateLogic.addUnseenThreadId(newMessageThreadId)
         }
-        // Update the thread inline if it is already loaded
-        stateLogic.upsertReply(reply = event.message)
+        // If the thread is loaded, it will be updated by message.new + message.updated events
     }
 
     /**
@@ -154,6 +200,30 @@ internal class QueryThreadsLogic(private val stateLogic: QueryThreadsStateLogic)
     }
 
     /**
+     * Marks a given thread as read by a user, if the [MessageReadEvent] is delivered for a thread.
+     *
+     * @param event The [MessageReadEvent] informing about the read state change.
+     */
+    private fun markThreadAsRead(event: MessageReadEvent) {
+        val threadInfo = event.thread ?: return
+        stateLogic.markThreadAsReadByUser(
+            threadInfo = threadInfo,
+            user = event.user,
+            createdAt = event.createdAt,
+        )
+    }
+
+    /**
+     * Marks a given thread as unread by a user, if the [NotificationMarkUnreadEvent] is delivered for a thread.
+     *
+     * @param event The [NotificationMarkUnreadEvent] informing about the read state change.
+     */
+    private fun markThreadAsUnread(event: NotificationMarkUnreadEvent) {
+        val threadId = event.threadId ?: return
+        stateLogic.markThreadAsUnreadByUser(threadId, event.user, event.createdAt)
+    }
+
+    /**
      * Deletes all threads associated with the channel with [cid].
      * Use when the channel was deleted.
      */
@@ -161,5 +231,13 @@ internal class QueryThreadsLogic(private val stateLogic: QueryThreadsStateLogic)
         val threads = stateLogic.getThreads()
         val filteredThreads = threads.filterNot { it.cid == cid }
         stateLogic.setThreads(filteredThreads)
+    }
+
+    private suspend fun queryThreadsOffline(request: QueryThreadsRequest) {
+        val localThreadsOrder = databaseLogic.getLocalThreadsOrder().take(request.limit)
+        if (localThreadsOrder.isEmpty()) return
+        val localThreads = databaseLogic.getLocalThreads(localThreadsOrder)
+        if (localThreads.isEmpty()) return
+        stateLogic.insertThreadsIfAbsent(localThreads)
     }
 }